+++ lib/widgets/dialogs/image_carousel_picker.dart
@@ -1,3 +1,4 @@
+import '../../utils/image_limits.dart' show boundedFileImage;
 import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';